Moved suppression interfaces to Core.
* no code changes
* move the suppression interfaces that don't depend on VS to Core
* moved interfaces are now in the namespace "SonarLint.VisualStudio.Core.Suppression"
* fixed up namespace references
