+++ src/Integration.Vsix/Suppression/LiveIssue.cs
@@ -19,7 +19,7 @@
  */
 
 using System.IO;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Suppression;
 using SonarLint.VisualStudio.Integration.Suppression;
 
 namespace SonarLint.VisualStudio.Integration.Vsix.Suppression
