+++ src/Integration/Suppression/SonarQubeIssuesProvider.cs
@@ -25,7 +25,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Suppression;
 using SonarLint.VisualStudio.Core.SystemAbstractions;
 using SonarQube.Client;
 using SonarQube.Client.Models;
