+++ src/Integration/Suppression/SuppressedIssueMatcher.cs
@@ -20,7 +20,7 @@
 
 using System;
 using System.Linq;
-using SonarLint.VisualStudio.Core;
+using SonarLint.VisualStudio.Core.Suppression;
 
 namespace SonarLint.VisualStudio.Integration.Suppression
 {
